VFS-292: debug

+++ include/bufferAgent.h
@@ -10,7 +10,7 @@
 #include "fileCache.h"
 
 typedef struct fuse_file_info*  ffi_type;
-typedef uint32_t                fd_type;
+typedef uint64_t                fd_type;
 
 typedef boost::function<int(std::string path, const std::string &buf, size_t, off_t, ffi_type)>    write_fun;
 typedef boost::function<int(std::string path, std::string &buf, size_t, off_t, ffi_type)>          read_fun;
